End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -54,16 +55,13 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
 
     }
-
+    [Authorize]
     [HttpPost]
-
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var aucton = _mapper.Map<Auction>(auctionDto);
 
-        // TO DO: add current user
-
-        aucton.Seller = "BoB martin";
+        aucton.Seller = User.Identity.Name;
 
         _context.Auctions.Add(aucton);
 
@@ -80,8 +78,8 @@ public class AuctionsController : ControllerBase
 
     }
 
+    [Authorize]
     [HttpPut("{id}")]
-
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
         var auction = await _context.Auctions
@@ -89,6 +87,8 @@ public class AuctionsController : ControllerBase
         .FirstOrDefaultAsync(x => x.Id == id);
 
         if (auction == null) return NotFound();
+
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -105,12 +105,14 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");
     }
 
-
+      [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id){
         var auction = await _context.Auctions.FindAsync(id);
         if (auction == null) return NotFound();
-        // TO DO : check the user
+        
+        if (auction.Seller != User.Identity.Name) return Forbid();
+
         _context.Auctions.Remove(auction);
         await _publishEndpoint.Publish<AuctionDteleted>(new { Id = auction.Id.ToString() });
         var result = await _context.SaveChangesAsync() > 0;
